% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocked_gibbs.R
\name{blocked_gibbs}
\alias{blocked_gibbs}
\title{Blocked Gibbs Sampler for Infinite Gaussian Mixture Model}
\usage{
blocked_gibbs(
  y,
  K = 20,
  lambda0 = 0.1,
  mu0 = 0,
  sigmasq0 = 10^2,
  alpha0 = 0.001,
  beta0 = 0.001,
  n_mcmc = 5000,
  burnin = 2500,
  n_message = 500
)
}
\arguments{
\item{n_message}{}
}
\value{

}
\description{
Blocked Gibbs Sampler for Infinite Gaussian Mixture Model
}
