% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{Title}
\usage{
mcmc(
  y,
  w,
  B = 2,
  K = 20,
  n_mcmc = 5000,
  burnin = 2500,
  n_message = 500,
  prior_parameters = NULL
)
}
\arguments{
\item{prior_parameters}{}
}
\value{

}
\description{
Title
}
