% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_old.R
\name{mcmc_old}
\alias{mcmc_old}
\title{Title}
\usage{
mcmc_old(
  y,
  w,
  B = 2,
  K = 20,
  lambda0 = 0.1,
  mu0 = 0,
  sigmasq0 = 10^2,
  alpha0 = 0.001,
  beta0 = 0.001,
  n_mcmc = 5000,
  burnin = 2500,
  n_message = 500
)
}
\arguments{
\item{n_message}{}
}
\value{

}
\description{
Title
}
