% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Data From Mixture Model}
\usage{
simulate_data(
  N = 100,
  K = 3,
  mu = c(-20, 5, 25),
  sigmasq = c(0.5, 0.5, 0.5),
  probs = c(1/3, 1/3, 1/3),
  plot = TRUE
)
}
\arguments{
\item{plot}{}
}
\value{

}
\description{
Simulate Data From Mixture Model
}
