% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mixture.R
\name{simulate_mixture}
\alias{simulate_mixture}
\title{Title}
\usage{
simulate_mixture(
  N = 500,
  B = 2,
  K = 3,
  mu = c(-20, 5, 25),
  sigmasq = c(0.5, 0.5, 0.5),
  phi = c(1/2, 1/2),
  p = list(c(1/3, 1/3, 1/3), c(1/2, 1/4, 1/4))
)
}
\arguments{
\item{p}{}
}
\value{

}
\description{
Title
}
