% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_breaking.R
\name{stick_breaking}
\alias{stick_breaking}
\title{Stick Breaking Representation for Dirichlet Process}
\usage{
stick_breaking(V, K)
}
\arguments{
\item{K}{}
}
\value{

}
\description{
Stick Breaking Representation for Dirichlet Process
}
