% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_breaking_expectation.R
\name{stick_breaking_expectation}
\alias{stick_breaking_expectation}
\title{Title}
\usage{
stick_breaking_expectation(B, K, alpha_q_V, beta_q_V)
}
\arguments{
\item{beta_q_V}{}
}
\value{

}
\description{
Title
}
