% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_expectations.R
\name{update_expectations}
\alias{update_expectations}
\title{Title}
\usage{
update_expectations(mod, variational_parameters = NULL, expectations = NULL)
}
\arguments{
\item{expectations}{}
}
\value{

}
\description{
Title
}
