% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_optimal_z.R
\name{update_optimal_z}
\alias{update_optimal_z}
\title{Title}
\usage{
update_optimal_z(y, N, K, E_log_pi, E_sigma_sq, E_log_sigma_sq, E_SS)
}
\arguments{
\item{E_SS}{}
}
\value{

}
\description{
Title
}
