% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_phi_old.R
\name{update_phi_old}
\alias{update_phi_old}
\title{Title}
\usage{
update_phi_old(
  y,
  N,
  B,
  K,
  E_log_phi,
  E_log_p,
  E_sigma_sq,
  E_log_sigma_sq,
  E_SS
)
}
\arguments{
\item{E_SS}{}
}
\value{

}
\description{
Title
}
