% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variational_parameters.R
\name{update_variational_parameters}
\alias{update_variational_parameters}
\title{Title}
\usage{
update_variational_parameters(
  mod,
  variational_parameters = NULL,
  expectations = NULL,
  prior_parameters = NULL,
  initialize = FALSE,
  initial_values = NULL
)
}
\arguments{
\item{initial_values}{}
}
\value{

}
\description{
Title
}
