% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBMIM.R
\name{VBMIM}
\alias{VBMIM}
\title{Title}
\usage{
VBMIM(
  y,
  B = 2,
  K = 20,
  mu0 = 0,
  sigmasq0 = 10^2,
  A0 = 0.01,
  B0 = 0.01,
  beta = 0.5,
  max_iter = 30
)
}
\arguments{
\item{max_iter}{}
}
\value{

}
\description{
Title
}
