% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbmm.R
\name{vbmm}
\alias{vbmm}
\title{Title}
\usage{
vbmm(
  y,
  B = 2,
  K = 20,
  max_iter = 30,
  w = NULL,
  prior_parameters = NULL,
  initial_values = NULL
)
}
\arguments{
\item{initial_values}{}
}
\value{

}
\description{
Title
}
