% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{vi}
\alias{vi}
\title{Variational Inference for Infinite Mixture Model}
\usage{
vi(
  y,
  K = 20,
  mu0 = 0,
  sigmasq0 = 10^8,
  A = 0.01,
  B = 0.01,
  beta = 0.5,
  max_iter = 30
)
}
\arguments{
\item{max_iter}{}
}
\value{

}
\description{
Variational Inference for Infinite Mixture Model
}
