% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick_breaking_expectation.R
\name{stick_breaking_expectation}
\alias{stick_breaking_expectation}
\title{Title}
\usage{
stick_breaking_expectation(B, K, E_log_V, E_log_1_V)
}
\arguments{
\item{E_log_1_V}{}
}
\value{

}
\description{
Title
}
