% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbmm.R
\name{vbmm}
\alias{vbmm}
\title{Title}
\usage{
vbmm(
  y,
  B = 2,
  K = 20,
  mu0 = 0,
  sigmasq0 = 10^8,
  A0 = 0.01,
  B0 = 0.01,
  alpha_phi0 = 1,
  beta_V0 = 0.5,
  max_iter = 30
)
}
\arguments{
\item{max_iter}{}
}
\value{

}
\description{
Title
}
